move auth directory under admin
