+++ apps/kita/composer.json
@@ -13,6 +13,7 @@
         "laravelcollective/html": "^6.4"
     },
     "require-dev": {
+        "barryvdh/laravel-debugbar": "^3.8",
         "fakerphp/faker": "^1.9.1",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.0.1",
