+++ apps/kita/routes/web.php
@@ -32,7 +32,7 @@ Route::get('/', function () {
 
 //admins middleware
 Route::group(['prefix' => 'admin', 'middleware' => ['auth:admins']], function () {
-    Route::post('/logout', [App\Http\Controllers\Admin\LoginController::class,'logout'])->name('admin.logout');
+    Route::post('/logout', [\App\Http\Controllers\Admin\Auth\LoginController::class,'logout'])->name('admin.logout');
     Route::get('/admin_users', [AdminController::class, 'index'])->name('admin_users.index');
     Route::get('/admin_users_create', [AdminController::class, 'create'])->name('admin_users.create');
     Route::post('/admin_users', [AdminController::class, 'store'])->name('admin_users.store');
@@ -51,8 +51,8 @@ Route::group(['prefix' => 'admin', 'middleware' => ['auth:admins']], function ()
 
 //admins middlewareなし
 Route::group(['prefix' => 'admin'], function () {
-    Route::get('/login', [App\Http\Controllers\Admin\LoginController::class, 'showLoginForm'])->name('admin.login.form');
-    Route::post('/login', [App\Http\Controllers\Admin\LoginController::class, 'login'])->name('admin.login');
+    Route::get('/login', [\App\Http\Controllers\Admin\Auth\LoginController::class, 'showLoginForm'])->name('admin.login.form');
+    Route::post('/login', [\App\Http\Controllers\Admin\Auth\LoginController::class, 'login'])->name('admin.login');
 });
 
 Route::get('/member_registration', [RegisterController::class, 'showRegistrationForm'])->name('register.form');
